feat: further refactor vol 2
